Storage settings.
