+++ src/NServiceBus.SqlServer/NServiceBus.SqlServer.csproj
@@ -103,6 +103,7 @@
     <Compile Include="Resharper.Annotations.g.cs" />
     <Compile Include="Sending\MessageDispatcher.cs" />
     <Compile Include="Configuration\SqlScopeOptions.cs" />
+    <Compile Include="Subscriptions\SubscriptionStoreSettings.cs" />
     <Compile Include="Subscriptions\SubscriptionReader.cs" />
     <Compile Include="SqlServerTransport.cs" />
     <Compile Include="Receiving\MessagePump.cs" />
